+++ lib/core/constant/colors_const.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/material.dart';
+
+class ConstColor {
+  static const Color primaryColor = Color(0xffF5A342);
+  static const Color redColor = Color(0xffFF0000);
+}
