+++ lib/feature/model/recipe_model.dart
@@ -1,20 +1,27 @@
+// ignore_for_file: public_member_api_docs, sort_constructors_first
 class RecipeModel {
   List<RecipeType> recipeTypes;
-
-  RecipeModel(this.recipeTypes);
+  RecipeModel({
+    required this.recipeTypes,
+  });
 }
 
 class RecipeType {
   String typeName;
   List<Recipe> recipes;
-
-  RecipeType(this.typeName, this.recipes);
+  RecipeType({
+    required this.typeName,
+    required this.recipes,
+  });
 }
 
 class Recipe {
   String name;
-  bool isLiked;
+  bool? isLiked;
   String image;
-
-  Recipe(this.name, this.isLiked, this.image);
+  Recipe({
+    required this.name,
+    this.isLiked,
+    required this.image,
+  });
 }
